refactor(persistence): move auto-analysis-constants into persistence/ and ratchet the allowlist
