+++ assistant/src/__tests__/persistence-layering-guard.test.ts
@@ -52,7 +52,6 @@ const MEMORY_DIR = join(ASSISTANT_SRC, "memory");
  */
 const PERSISTENCE_TO_MEMORY_ALLOWLIST: Record<string, ReadonlySet<string>> = {
   "assistant/src/persistence/conversation-crud.ts": new Set([
-    "auto-analysis-constants",
     "graph/graph-memory-state-store",
     "indexer",
     "memory-retrospective-constants",
