+++ assistant/src/persistence/auto-analysis-constants.ts
